THERE ARE FOUR LIGHTS
as part of the deal, I needed to change the player tick time
so we can get it with a timer overflow. now instead of 10ms,
it is 8.192ms. songs previously converted will play fast.

also, both timers are doing LED PWM. song playing is moved
out of the interrupt handler, which it never should have been in
to begin with. I've tried to deal with the player code getting
behind but I think there are still a few issues.